LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-animation-corner-cases.tentative-expected.txt


FAIL Descendent animations should keep the pop up visible until the animation ends assert_equals: There should be two animations running expected 2 but got 0
PASS Pre-existing animations should *not* keep the pop up visible until the animation ends
FAIL It should be possible to use the "beforetoggle" event handler to animate the hide assert_equals: the hide animation should now be running expected 1 but got 0
FAIL It should be possible to use the "beforetoggle" event handler to animate the hide, even when the hide is due to dialog.showModal assert_equals: the hide animation should now be running expected 1 but got 0
PASS toggle event cannot be cancelled
FAIL Closing animations are triggered by changing the popover type assert_equals: There should be two animations running expected 2 but got 0
FAIL animation finish/cancel events must be trusted in order to finish closing the popover. assert_equals: the hide animation should now be running expected 1 but got 0
FAIL Capturing event listeners can't affect popover animations. assert_true: The popover should still be visible because the animation hasn't ended. expected true got false
